Initial support for multiple files.
This commit adds preliminary support for more then
one file using the --file option.
It defaults to todo.txt, but any amount of files can
be provided in the config file.

Note that this requires a change of setup in the
 config file, which is not backwards compatible!
 hence we set a DEPRECATION warning if old config
 is found.

This is the first step towards more actions, such
as

* "move" to move todos between files,
* "archive" to move done items to done.txt.
* "backup" to move files to a "backup.txt" and so on.

For now, however, one can simply add --file=foo to perform
any todotxt action on that file instead of the default todo.txt.